Refatorando a model, movendo taxas diarias para classe Projeto
